fix: 홈 경로 수정
